+++ src/main/java/com/shoppingplatform/price/api/dto/ProductDto.java
@@ -3,6 +3,8 @@ package com.shoppingplatform.price.api.dto;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -12,5 +14,5 @@ public class ProductDto {
 
     @NotNull(message = "Price cannot be null")
     @Min(value = 0, message = "Price must be greater than or equal to 0")
-    private Double price;
+    private BigDecimal price;
 }
